requirements.txt
Django~>2.1.2
django-summernote==0.8.10.0
djangorestframework==3.8.2
pytz==2018.5
